+++ app/src/main/java/com/example/orders/MainActivity.java
@@ -1,10 +1,13 @@
 package com.example.orders;
 
+import static com.example.orders.R.string.cancel;
+
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -48,15 +51,27 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void deleteOrder(View view) {
-        try {
-            dataBaseWorker.deleteOrders(true, "");
-            orders = dataBaseWorker.loadData();
-            cardAdapter.notifyDataSetChanged();
-            finish();
-            startActivity(getIntent());
-        } catch (Exception e) {
-            Toast.makeText(this, e.getMessage(),
-                    Toast.LENGTH_SHORT).show();
-        }
+        new AlertDialog.Builder(view.getContext())
+                .setTitle(R.string.deletion)
+                .setMessage(R.string.delete_message_all)
+                .setPositiveButton(R.string.delete, (dialog, which) -> {
+                    try {
+                        dataBaseWorker.deleteOrders(true, "");
+                        orders = dataBaseWorker.loadData();
+                        cardAdapter.notifyDataSetChanged();
+                        finish();
+                        startActivity(getIntent());
+                    } catch (Exception e) {
+                        Toast.makeText(this, e.getMessage(),
+                                Toast.LENGTH_SHORT).show();
+                    }
+                })
+                .setNegativeButton(R.string.cancel, (dialog, which) -> {
+                    Toast.makeText(view.getContext(), cancel,
+                            Toast.LENGTH_SHORT).show();
+                })
+                .setIcon(android.R.drawable.ic_dialog_alert)
+                .show();
+
     }
 }
